Refactor Mujoco and USD model factories to separate  visual builders from model ones
