+++ src/adam/model/mj_factory/mujoco_model.py
@@ -6,18 +6,12 @@ import numpy as np
 from scipy.spatial.transform import Rotation as R
 
 from adam.core.spatial_math import SpatialMath
-from adam.model.abc_factories import Limits, ModelFactory, Pose
-from adam.model.visuals import (
-    BoxVisualGeometry,
-    CylinderVisualGeometry,
-    EmbeddedMeshVisualGeometry,
-    SphereVisualGeometry,
-    Visual,
-    VisualMaterial,
-    capsule_mesh_geometry,
-)
+from adam.model.abc_factories import Limits, ModelFactory
+from adam.model.visuals import Visual
 from adam.model.std_factories.std_joint import StdJoint
 from adam.model.std_factories.std_link import StdLink
+
+from .mujoco_visual import MujocoVisualBuilder
 
 # Type checking only - doesn't execute at runtime
 if TYPE_CHECKING:
@@ -92,6 +86,13 @@ class MujocoModelFactory(ModelFactory):
         self.mj_model = self._model_exists(mj_model)
         fallback_name = "mujoco_model"
         self.name = getattr(self.mj_model, "name", None) or fallback_name
+        self._visual_builder = MujocoVisualBuilder(
+            mujoco=self.mujoco,
+            mj_model=self.mj_model,
+            math=self.math,
+            normalize_quaternion=_normalize_quaternion,
+            quat_to_rpy=_quat_to_rpy,
+        )
 
         self._links = self._build_links()
         self._child_map = self._build_child_map()
@@ -149,110 +150,8 @@ class MujocoModelFactory(ModelFactory):
         )
         return MujocoInertial(mass=mass, inertia=inertia, origin=origin)
 
-    def _geom_name(self, geom_id: int) -> str:
-        name = self.mujoco.mj_id2name(
-            self.mj_model, self.mujoco.mjtObj.mjOBJ_GEOM, geom_id
-        )
-        return name if name is not None else f"geom_{geom_id}"
-
-    def _pose_from_pos_quat(self, pos: np.ndarray, quat: np.ndarray) -> Pose:
-        return Pose.build(np.asarray(pos, dtype=float), _quat_to_rpy(quat), self.math)
-
-    def _compiled_mesh_geometry(self, mesh_id: int) -> EmbeddedMeshVisualGeometry:
-        vert_start = int(self.mj_model.mesh_vertadr[mesh_id])
-        vert_count = int(self.mj_model.mesh_vertnum[mesh_id])
-        face_start = int(self.mj_model.mesh_faceadr[mesh_id])
-        face_count = int(self.mj_model.mesh_facenum[mesh_id])
-        return EmbeddedMeshVisualGeometry(
-            vertices=np.asarray(
-                self.mj_model.mesh_vert[vert_start : vert_start + vert_count],
-                dtype=np.float32,
-            ).copy(),
-            faces=np.asarray(
-                self.mj_model.mesh_face[face_start : face_start + face_count],
-                dtype=np.uint32,
-            ).copy(),
-        )
-
-    def _geom_visual(self, geom_id: int) -> Visual | None:
-        geom_type = int(self.mj_model.geom_type[geom_id])
-        geom_size = np.asarray(self.mj_model.geom_size[geom_id], dtype=float)
-        geom_pos = np.asarray(self.mj_model.geom_pos[geom_id], dtype=float)
-        geom_quat = _normalize_quaternion(
-            np.asarray(self.mj_model.geom_quat[geom_id], dtype=float)
-        )
-        material = VisualMaterial(
-            rgba=tuple(float(value) for value in self.mj_model.geom_rgba[geom_id])
-        )
-
-        if geom_type == self.mujoco.mjtGeom.mjGEOM_BOX:
-            geometry = BoxVisualGeometry(
-                size=tuple(float(2.0 * v) for v in geom_size[:3])
-            )
-            origin = self._pose_from_pos_quat(geom_pos, geom_quat)
-        elif geom_type == self.mujoco.mjtGeom.mjGEOM_SPHERE:
-            geometry = SphereVisualGeometry(radius=float(geom_size[0]))
-            origin = self._pose_from_pos_quat(geom_pos, geom_quat)
-        elif geom_type == self.mujoco.mjtGeom.mjGEOM_CYLINDER:
-            geometry = CylinderVisualGeometry(
-                radius=float(geom_size[0]),
-                length=float(2.0 * geom_size[1]),
-            )
-            origin = self._pose_from_pos_quat(geom_pos, geom_quat)
-        elif geom_type == self.mujoco.mjtGeom.mjGEOM_CAPSULE:
-            geometry = capsule_mesh_geometry(
-                radius=float(geom_size[0]),
-                cylindrical_length=float(2.0 * geom_size[1]),
-            )
-            origin = self._pose_from_pos_quat(geom_pos, geom_quat)
-        elif geom_type == self.mujoco.mjtGeom.mjGEOM_MESH:
-            mesh_id = int(self.mj_model.geom_dataid[geom_id])
-            geometry = self._compiled_mesh_geometry(mesh_id)
-            origin = self._pose_from_pos_quat(geom_pos, geom_quat)
-        else:
-            return None
-
-        return Visual(
-            name=self._geom_name(geom_id),
-            origin=origin,
-            geometry=geometry,
-            material=material,
-        )
-
-    def _geom_signature(self, geom_id: int) -> tuple:
-        return (
-            int(self.mj_model.geom_type[geom_id]),
-            int(self.mj_model.geom_dataid[geom_id]),
-            tuple(float(v) for v in self.mj_model.geom_size[geom_id]),
-            tuple(float(v) for v in self.mj_model.geom_pos[geom_id]),
-            tuple(float(v) for v in self.mj_model.geom_quat[geom_id]),
-        )
-
     def _link_visuals(self, body_id: int) -> list[Visual]:
-        visuals: list[Visual] = []
-        geom_ids = [
-            geom_id
-            for geom_id, geom_body_id in enumerate(self.mj_model.geom_bodyid)
-            if int(geom_body_id) == body_id
-        ]
-        non_contact_signatures = {
-            self._geom_signature(geom_id)
-            for geom_id in geom_ids
-            if int(self.mj_model.geom_contype[geom_id]) == 0
-            and int(self.mj_model.geom_conaffinity[geom_id]) == 0
-        }
-
-        for geom_id in geom_ids:
-            signature = self._geom_signature(geom_id)
-            if (
-                int(self.mj_model.geom_contype[geom_id]) != 0
-                or int(self.mj_model.geom_conaffinity[geom_id]) != 0
-            ) and signature in non_contact_signatures:
-                continue
-            visual = self._geom_visual(geom_id)
-            if visual is not None:
-                visuals.append(visual)
-        return visuals
+        return self._visual_builder.link_visuals(body_id)
 
     def _build_links(self) -> list[StdLink]:
         links: list[StdLink] = []
